patch-25.74m: modularize layout engine
